Enumerate.Function.Extra

Documentation

nothing :: Monad m => m ()

maybe2bool :: Maybe a -> Bool

either2maybe :: Either e a -> Maybe a

either2bool :: Either e a -> Bool

failed :: MonadThrow m => String -> m a

failed = throwM . userError

maybe2throw :: (a -> Maybe b) -> forall m. MonadThrow m => a -> m b

generalize a function that fails with Nothing.

list2throw :: (a -> [b]) -> forall m. MonadThrow m => a -> m b

generalize a function that fails with [].

either2throw :: (a -> Either SomeException b) -> forall m. MonadThrow m => a -> m b

generalize a function that fails with Left.

throw2maybe :: (forall m. MonadThrow m => a -> m b) -> a -> Maybe b

specialization

throw2either :: (forall m. MonadThrow m => a -> m b) -> a -> Either SomeException b

specialization

throw2list :: (forall m. MonadThrow m => a -> m b) -> a -> [b]

specialization

totalizeFunction :: (NFData b, MonadThrow m) => (a -> b) -> a -> m b

makes an *unsafely*-partial function (i.e. a function that throws exceptions or that has inexhaustive pattern matching) into a *safely*-partial function (i.e. that explicitly returns in a monad that supports failure).

defaultPartialityHandlers :: MonadThrow m => [Handler (m a)]

handles the following exceptions:

ArithException

ArrayException

ErrorCall

PatternMatchFail

spoonWith :: (NFData a, MonadThrow m) => [Handler (m a)] -> a -> m a

Evaluate a value to normal form and throwM any exceptions are thrown during evaluation. For any error-free value, spoon = Just.

(taken from the spoon package.)

showsPrecWith :: Show b => String -> (a -> b) -> Int -> a -> ShowS

the eliminator as a function and the introducer as a string

helper for declaring Show instances of datatypes without visible constructors (like Map which is shown as a list).

module Control.DeepSeq

module Data.Semigroup

module GHC.Generics

module Data.Data

module Control.Arrow

module Data.Function

module Data.List

module Data.Foldable